Move error module a level up
Makes it available for resources as well
