+++ src/terminal/dumb.rs
@@ -17,8 +17,8 @@
 use failure::Error;
 use std::io::Write;
 
+use super::super::error::NotSupportedError;
 use super::super::resources::{Resource, ResourceAccess};
-use super::error::NotSupportedError;
 use super::size::Size;
 use super::write::Terminal;
